Fix price printout

+++ app/src/main/java/com/wen/database/model/Menu_Item.java
@@ -1,5 +1,7 @@
 package com.wen.database.model;
 
+import java.math.BigDecimal;
+
 /**
  * Created by Wen on 7/27/2015.
  */
@@ -7,13 +9,13 @@ public class Menu_Item {
     private Integer id;
     private String name;
     private String description;
-    private Double price;
+    private BigDecimal price;
     private Category category;
 
     public Menu_Item() {
     }
 
-    public Menu_Item(Integer id, String name, String description, Double price, Category category) {
+    public Menu_Item(Integer id, String name, String description, BigDecimal price, Category category) {
         this.id = id;
         this.name = name;
         this.description = description;
@@ -45,11 +47,11 @@ public class Menu_Item {
         this.description = description;
     }
 
-    public Double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(Double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
